Update view with tabs
