+++ plugin-flex-ts-template-v2/src/feature-library/contacts/custom-components/ContactsSideLink.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { SideLink, Actions, Manager } from '@twilio/flex-ui';
 
-import { StringTemplates } from '../strings';
+import { StringTemplates } from '../flex-hooks/strings';
 
 interface OwnProps {
   activeView?: string;
