+++ plugin-flex-ts-template-v2/src/feature-library/contacts/flex-hooks/components/SideNav.tsx
@@ -2,13 +2,13 @@ import * as Flex from '@twilio/flex-ui';
 
 import { FlexComponent } from '../../../../types/feature-loader';
 import { isFeatureEnabled } from '../../config';
-import ContactHistorySideLink from './ContactHistorySideLink';
+import ContactsSideLink from '../../custom-components/ContactsSideLink';
 
 export const componentName = FlexComponent.SideNav;
-export const componentHook = function addContactHistoryToSideNav(flex: typeof Flex) {
+export const componentHook = function addContactsToSideNav(flex: typeof Flex) {
   if (!isFeatureEnabled()) {
     return;
   }
 
-  flex.SideNav.Content.add(<ContactHistorySideLink viewName="contact-history" key="contact-history-side-nav" />);
+  flex.SideNav.Content.add(<ContactsSideLink viewName="contacts" key="contacts-side-nav" />);
 };
