+++ plugin-flex-ts-template-v2/src/feature-library/contacts/flex-hooks/components/ViewCollection.tsx
@@ -1,13 +1,13 @@
 import * as Flex from '@twilio/flex-ui';
 
 import { FlexComponent } from '../../../../types/feature-loader';
-import ContactHistory from '../../custom-components/ContactHistoryView/ContactHistoryView';
+import ContactsView from '../../custom-components/ContactsView/ContactsView';
 
 export const componentName = FlexComponent.ViewCollection;
-export const componentHook = function addContactHistoryView(flex: typeof Flex) {
+export const componentHook = function addContactsView(flex: typeof Flex) {
   flex.ViewCollection.Content.add(
-    <flex.View name="contact-history" key="contact-history-view">
-      <ContactHistory key="contact-history-view-content" />
+    <flex.View name="contacts" key="contacts-view">
+      <ContactsView key="contacts-view-content" />
     </flex.View>,
   );
 };
