+++ plugin-flex-ts-template-v2/src/feature-library/contacts/flex-hooks/paste-elements/index.ts
@@ -4,4 +4,9 @@ export const pasteElementHook = {
   COMPACT_MODAL: {
     width: '400px',
   },
+  CONTACTS_VIEW_WRAPPER: {
+    height: '100%',
+    padding: 'space40',
+    overflow: 'auto',
+  },
 } as { [key: string]: PasteCustomCSS };
